removed bootstrap, and some inpage style overrides
removed bootstrap from base.  We were not using it much
and it was causing some unpredictability with certain classes that might have
shared a class name with bootstrap.

removed inpage styling from index

added some of the magic that was going on in bootstrap to our styles.css
headers font weight being set to 500

along with button font and size.