bugfixes, dynamic tags, flappybird cheat check, stop duplicate crawls
crawl2 now checks if a game with the title already exists and skips if so

pagination included now on index route, same with template

cheat check using session timestamps of start and end, and if the score falls within an acceptable range.

style changes for pagination

game route passes tags to the page and game page now iterates through buttons found

searchbar made a note about a div not being used.
searchbar js updated for the correct cover location.