moved game catalogue to a partial
Super frustrating "feature" with chrome that would cause the page to hang forever,
because each call to a next tab was a new call to the index page.

chrome did not like the video banner getting called each time.

created ajax style game catalogue so that the page does not reload when tabbing through.